+++ docs/Changelog.md
@@ -2,6 +2,23 @@
 
 # ONNX-TensorRT Changelog
 
+## TensorRT 8.2 EA Release - 2021-10-04
+### Added
+- Added support for the following ONNX operators:
+  - Einsum
+  - IsNan
+  - GatherND
+  - Scatter
+  - ScatterElements
+  - ScatterND
+  - Sign
+  - Round
+
+### Updated
+- Updated `Gather` and `GatherElements` implementations to natively support negative indices
+- Updated `Pad` layer to support ND padding, along with `edge` and `reflect` padding mode support
+- Updated `If` layer with general performance improvements.
+
 ## TensorRT 8.0 Release - 2021-07-02
 ### Added
  - Rehauled resize operator, now fully supporting the following modes: